Groundtruth stored in test set entry
After consulting existing project test sets we come to a conclusion
that existing tests sets has different inputs for different groundtruths
(even in the same test set). As a result our assumption of test set
entry and groundtruth models was not correct. This was refactored and
right now input and groundtruth are stored in the `TestSetEntry` model.