Remove duplicated relation between task and test set
Since `input` and `groundtruth` are located in `TestSetEntry` model we
also needed to connect this model with the task (to know how to convert
input to hipothese). As a result we faced the problem with duplicated
relation between task and test set. Here this situation is solved by
removed direct relation between task and test set (done though
`TaskTestSet` model).