Rest API for submitting evaluation scores
Evaluation scores REST API is located at:

```
POST /evaluations/:evaluation_id/scores
```

The request body should contain all metrics scores (when any of the
metrics scores is missing, 422 error code is returned) in the following
format:

```json
{ "scores": {
    "metric1_name": 1,
    "metric2_name": 2.2,
    ...
  }
}
```

API is secured by the evaluation token given as a bearer value. Token
can be regenerated on demand:

```ruby
evaluation = Evaluation.first
token = evaluation.reset_token!
```

and this is the only time the value is returned (the token is stored in
the DB as a secure hash).