Fix issue with updating job status
slurmrestd is returning job id as a number and we are storring it in our
DB as string. As a result existing jobs statuses were set to `failed`,
since we are treating not found jobs in slurm current db as finished and
when job is finished without scores it is treated as failed job.

The fix was to convert statuses job ids to string.